Finished the setup for 3d-Widget and it works just like the HUD and add to view
had to come up with a new class WidgetContainerActor and this class is just like the HUD and have all the functions and some more because it need a different way to set the widget instead of just AddtoViewPort.
one key thing was that VRPAwn in the default unreal VR template is set up completely with blueprint and also to keep the modal structure of the game and not hardcoding I made the AbilitySystemComponent BlueprintReadOnly and added the child actor component of the BP_WidgetContainerActor with in the blueprint and call the OnPossess on blueprint to replicate what I did in VRPawnBase inside of the blueprint.
I tested in VR mode and it worked fine just have to change it size to fit VR.